+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog
 
  - Catch error in grid gallery element if grid does not exist
  - Fix broken gallery image sorting
+ - Fix grid order in the options
 
 2.0.0 (2018-01-05)
 ------------------